move calculation state
