+++ source/dotnet/wholesale-api/Batches/Batches.Application/ICalculationInfrastructureService.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Energinet.DataHub.Wholesale.Batches.Application.Model;
 using Energinet.DataHub.Wholesale.Batches.Application.Model.Batches;
 
 namespace Energinet.DataHub.Wholesale.Batches.Application;
