+++ source/dotnet/wholesale-api/Batches/Batches.Application/Model/CalculationState.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Wholesale.Batches.Application;
+namespace Energinet.DataHub.Wholesale.Batches.Application.Model;
 
 public enum CalculationState
 {
