+++ source/dotnet/wholesale-api/Batches/Batches.Infrastructure/Calculations/CalculationInfrastructureService.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Batches.Application;
+using Energinet.DataHub.Wholesale.Batches.Application.Model;
 using Energinet.DataHub.Wholesale.Batches.Application.Model.Batches;
 
 namespace Energinet.DataHub.Wholesale.Batches.Infrastructure.Calculations;
