+++ source/dotnet/wholesale-api/Batches/Batches.UnitTests/Infrastructure/BatchExecutionStateDomainService/BatchStateMapperTests.cs
@@ -14,6 +14,7 @@
 
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Energinet.DataHub.Wholesale.Batches.Application;
+using Energinet.DataHub.Wholesale.Batches.Application.Model;
 using Energinet.DataHub.Wholesale.Batches.Application.Model.Batches;
 using Energinet.DataHub.Wholesale.Batches.Infrastructure.BatchState;
 using FluentAssertions;
